// types/dtos/ShopDto.ts
export interface ShopDto {
  id: string;
  name: string;
  address: string;
  isMainShop: boolean;
  countryIsoCode: string;
  countryName: string;
  provinceCode: string;
  provinceName: string;
  cityCode: string;
  cityName: string;
  subDistrictCode: string;
  subDistrictName: string;
  villageCode: string;
  villageName: string;
  postalCode: string;
}